// src/app/api/ui/enums/chat-message-type-alt.enum.ts
export enum ChatMessageTypeAltEnum {
  Undefined = 0,
  SendContractFromCustomerToSupplier = 1,
  ChangeContractByCustomer = 2,
  ChangeContractBySupplier = 3,
  DifferenceProtocol = 4,
  Simple = 5,
  ReworkContract = 6,
  SendContractFromSupplierToCustomer = 7,
  ChangeFormToPaperDeal = 8,
  SupplierAddDealProvision = 9,
  ReturnContractToForming = 10,
  ApplicationClarificationRequest = 11,
  ApplicationClarificationResponse = 12,
}